requirements.txt
Flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
xgboost==2.0.3
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
prometheus-flask-exporter==0.23.0
pydantic==2.5.3
pytest==7.4.3
requests==2.31.0